Move scope change events to the root level
