+++ src/components/Assets/AssetList/index.tsx
@@ -7,9 +7,9 @@ import { useAssetsSelector } from "../../../store/assets/useAssetsSelector";
 import { useConfigSelector } from "../../../store/config/useConfigSelector";
 import { useStore } from "../../../store/useStore";
 import { isEnvironment } from "../../../typeGuards/isEnvironment";
+import { SCOPE_CHANGE_EVENTS } from "../../../types";
 import { changeScope } from "../../../utils/actions/changeScope";
 import { sendUserActionTrackingEvent } from "../../../utils/actions/sendUserActionTrackingEvent";
-import { SCOPE_CHANGE_EVENTS } from "../../Main/types";
 import { EmptyState } from "../../common/EmptyState";
 import { Menu } from "../../common/Menu";
 import { NewCircleLoader } from "../../common/NewCircleLoader";
