+++ src/components/Assets/AssetTypeList/index.tsx
@@ -8,9 +8,9 @@ import { useStore } from "../../../store/useStore";
 import { isEnvironment } from "../../../typeGuards/isEnvironment";
 import { isNull } from "../../../typeGuards/isNull";
 import { isString } from "../../../typeGuards/isString";
+import { SCOPE_CHANGE_EVENTS } from "../../../types";
 import { changeScope } from "../../../utils/actions/changeScope";
 import { sendUserActionTrackingEvent } from "../../../utils/actions/sendUserActionTrackingEvent";
-import { SCOPE_CHANGE_EVENTS } from "../../Main/types";
 import { ChildIcon } from "../../common/icons/30px/ChildIcon";
 import { AssetFilterQuery } from "../AssetsFilter/types";
 import { ViewMode } from "../AssetsViewScopeConfiguration/types";
