+++ src/components/Dashboard/NewReport/index.tsx
@@ -2,10 +2,9 @@ import { useLayoutEffect, useMemo, useState } from "react";
 import { getFeatureFlagValue } from "../../../featureFlags";
 import { useConfigSelector } from "../../../store/config/useConfigSelector";
 import { isNull } from "../../../typeGuards/isNull";
-import { FeatureFlag } from "../../../types";
+import { FeatureFlag, SCOPE_CHANGE_EVENTS } from "../../../types";
 import { changeScope } from "../../../utils/actions/changeScope";
 import { DigmaLogoIcon } from "../../common/icons/16px/DigmaLogoIcon";
-import { SCOPE_CHANGE_EVENTS } from "../../Main/types";
 import { actions } from "../actions";
 import { Chart } from "./Chart";
 import { EmptyState } from "./EmptyState";
