+++ src/components/Highlights/Impact/index.tsx
@@ -2,9 +2,9 @@ import { Row, createColumnHelper } from "@tanstack/react-table";
 import { useEffect } from "react";
 import { PERFORMANCE_IMPACT_DOCUMENTATION_URL } from "../../../constants";
 import { useConfigSelector } from "../../../store/config/useConfigSelector";
+import { SCOPE_CHANGE_EVENTS } from "../../../types";
 import { openURLInDefaultBrowser } from "../../../utils/actions/openURLInDefaultBrowser";
 import { sendUserActionTrackingEvent } from "../../../utils/actions/sendUserActionTrackingEvent";
-import { SCOPE_CHANGE_EVENTS } from "../../Main/types";
 import { InfinityIcon } from "../../common/icons/16px/InfinityIcon";
 import { RefreshIcon } from "../../common/icons/16px/RefreshIcon";
 import { TargetIcon } from "../../common/icons/16px/TargetIcon";
