+++ src/components/Highlights/Performance/index.tsx
@@ -3,10 +3,10 @@ import { useEffect, useState } from "react";
 import { usePrevious } from "../../../hooks/usePrevious";
 import { useConfigSelector } from "../../../store/config/useConfigSelector";
 import { isBoolean } from "../../../typeGuards/isBoolean";
+import { SCOPE_CHANGE_EVENTS } from "../../../types";
 import { sendUserActionTrackingEvent } from "../../../utils/actions/sendUserActionTrackingEvent";
 import { formatTimeDistance } from "../../../utils/formatTimeDistance";
 import { getDurationString } from "../../../utils/getDurationString";
-import { SCOPE_CHANGE_EVENTS } from "../../Main/types";
 import { CrossCircleIcon } from "../../common/icons/16px/CrossCircleIcon";
 import { RefreshIcon } from "../../common/icons/16px/RefreshIcon";
 import { TimerIcon } from "../../common/icons/16px/TimerIcon";
