+++ src/components/Highlights/Scaling/index.tsx
@@ -2,11 +2,11 @@ import { Row, createColumnHelper } from "@tanstack/react-table";
 import { useEffect } from "react";
 import { SCALING_ISSUE_DOCUMENTATION_URL } from "../../../constants";
 import { useConfigSelector } from "../../../store/config/useConfigSelector";
+import { SCOPE_CHANGE_EVENTS } from "../../../types";
 import { openURLInDefaultBrowser } from "../../../utils/actions/openURLInDefaultBrowser";
 import { sendUserActionTrackingEvent } from "../../../utils/actions/sendUserActionTrackingEvent";
 import { getDurationString } from "../../../utils/getDurationString";
 import { InsightStatus } from "../../Insights/types";
-import { SCOPE_CHANGE_EVENTS } from "../../Main/types";
 import { useHistory } from "../../Main/useHistory";
 import { TAB_IDS } from "../../Navigation/Tabs/types";
 import { CrossCircleIcon } from "../../common/icons/16px/CrossCircleIcon";
