+++ src/components/Highlights/TopIssues/highlightCards/SpanQueryOptimizationHighlightCard/index.tsx
@@ -1,9 +1,9 @@
 import { Row, createColumnHelper } from "@tanstack/react-table";
 import { Duration } from "../../../../../globals";
 import { useConfigSelector } from "../../../../../store/config/useConfigSelector";
+import { SCOPE_CHANGE_EVENTS } from "../../../../../types";
 import { sendUserActionTrackingEvent } from "../../../../../utils/actions/sendUserActionTrackingEvent";
 import { getDurationString } from "../../../../../utils/getDurationString";
-import { SCOPE_CHANGE_EVENTS } from "../../../../Main/types";
 import { Tag } from "../../../../common/v3/Tag";
 import { Table } from "../../../common/Table";
 import { TableText } from "../../../common/TableText";
