+++ src/components/Insights/InsightsCatalog/InsightsPage/index.tsx
@@ -5,10 +5,9 @@ import { useConfigSelector } from "../../../../store/config/useConfigSelector";
 import { trackingEvents as globalEvents } from "../../../../trackingEvents";
 import { isNumber } from "../../../../typeGuards/isNumber";
 import { isUndefined } from "../../../../typeGuards/isUndefined";
-import { InsightType } from "../../../../types";
+import { InsightType, SCOPE_CHANGE_EVENTS } from "../../../../types";
 import { changeScope } from "../../../../utils/actions/changeScope";
 import { sendUserActionTrackingEvent } from "../../../../utils/actions/sendUserActionTrackingEvent";
-import { SCOPE_CHANGE_EVENTS } from "../../../Main/types";
 import { useHistory } from "../../../Main/useHistory";
 import { TAB_IDS } from "../../../Navigation/Tabs/types";
 import { Scope } from "../../../common/App/types";
