+++ src/components/Navigation/HistoryNavigationPanel/index.tsx
@@ -4,13 +4,10 @@ import { useTheme } from "styled-components";
 import { history } from "../../../containers/Main/history";
 import { HistoryEntry } from "../../../history/History";
 import { useConfigSelector } from "../../../store/config/useConfigSelector";
+import { SCOPE_CHANGE_EVENTS } from "../../../types";
 import { changeScope } from "../../../utils/actions/changeScope";
 import { sendUserActionTrackingEvent } from "../../../utils/actions/sendUserActionTrackingEvent";
-import {
-  HistoryState,
-  ReactRouterLocationState,
-  SCOPE_CHANGE_EVENTS
-} from "../../Main/types";
+import { HistoryState, ReactRouterLocationState } from "../../Main/types";
 import { useBrowserLocationUpdater } from "../../Main/updateBrowserLocationUpdater";
 import { useHistory } from "../../Main/useHistory";
 import { HomeIcon } from "../../common/icons/16px/HomeIcon";
