+++ src/components/Notifications/index.tsx
@@ -3,10 +3,10 @@ import { dispatcher } from "../../dispatcher";
 import { usePrevious } from "../../hooks/usePrevious";
 import { isBoolean } from "../../typeGuards/isBoolean";
 import { isNumber } from "../../typeGuards/isNumber";
+import { SCOPE_CHANGE_EVENTS } from "../../types";
 import { changeScope } from "../../utils/actions/changeScope";
 import { sendTrackingEvent } from "../../utils/actions/sendTrackingEvent";
 import { sendUserActionTrackingEvent } from "../../utils/actions/sendUserActionTrackingEvent";
-import { SCOPE_CHANGE_EVENTS } from "../Main/types";
 import { FullView } from "./FullView";
 import { RecentView } from "./RecentView";
 import { actions } from "./actions";
